examples/event_loop_detection/pyproject.toml
[project]
name = "pdf-ingest"
version = "0.1.0"
description = "Demo: Event loop blocking detection with pyleak"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.25",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "matplotlib>=3.10.0",
    "pydantic>=2.12.5",
    "pyleak>=0.1.14",
    "pymupdf>=1.26.7",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pillow>=11.2.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]

[tool.uv.sources]
pdf-ingest = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_ingest"]
